refactor(annotation): move ExportResult from core to api layer
