+++ src/pragmata/annotation/__init__.py
@@ -1,6 +1,9 @@
 """Public annotation namespace — re-exports from internal modules."""
 
 from pragmata.api.annotation_export import (
+    ExportResult as ExportResult,
+)
+from pragmata.api.annotation_export import (
     export_annotations as export_annotations,
 )
 from pragmata.api.annotation_import import (
@@ -14,9 +17,6 @@ from pragmata.api.annotation_setup import (
 )
 from pragmata.api.annotation_setup import (
     teardown as teardown,
-)
-from pragmata.core.annotation.export_helpers import (
-    ExportResult as ExportResult,
 )
 from pragmata.core.annotation.setup import (
     SetupResult as SetupResult,
